requirements.txt
# common
discord.py
pymongo

# stable only
pymongo[srv]
Flask
